Update gridengine_fromcomponentimage.json
removed the VM reference within the Network Security Group component of the network Interface resource. I'm getting errors within Azure that the account is not authorized to do a join action on the network security group. The error is badly defined so I'm having to hunt-and-peck to figure out where the problem really is.