Update gridengine_fromcomponentimage.json
I had the name of the storage account set to the Unique VM Name, which had too many characters to be accepted by Azure. I've updated my code to extract the unique numeric portion of the VM Name and use that as the storage account name. This should provide a unique name for the storage account and allow it to be included in the resource group for deletion purposes.